AI-Assessment/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

sqlalchemy==2.0.29
pymysql==1.1.0

python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

pydantic==2.6.4

python-multipart==0.0.9

sentence-transformers
faiss-cpu
pypdf
numpy